+++ src/new-components/text-field/text-field.tsx
@@ -1,17 +1,19 @@
 import * as React from 'react'
-import { BaseField } from '../base-field'
+import { BaseField, BaseFieldVariantProps } from '../base-field'
 import { Box } from '../box'
 import styles from './text-field.module.css'
 import type { FieldComponentProps } from '../base-field'
 
 type TextFieldType = 'email' | 'search' | 'tel' | 'text' | 'url'
 
-type TextFieldProps = Omit<FieldComponentProps<HTMLInputElement>, 'type'> & {
-    type?: TextFieldType
-}
+type TextFieldProps = Omit<FieldComponentProps<HTMLInputElement>, 'type'> &
+    BaseFieldVariantProps & {
+        type?: TextFieldType
+    }
 
 const TextField = React.forwardRef<HTMLInputElement, TextFieldProps>(function TextField(
     {
+        variant = 'default',
         id,
         label,
         secondaryLabel,
@@ -29,6 +31,7 @@ const TextField = React.forwardRef<HTMLInputElement, TextFieldProps>(function Te
 ) {
     return (
         <BaseField
+            variant={variant}
             id={id}
             label={label}
             secondaryLabel={secondaryLabel}
@@ -41,7 +44,13 @@ const TextField = React.forwardRef<HTMLInputElement, TextFieldProps>(function Te
             aria-describedby={ariaDescribedBy}
         >
             {(extraProps) => (
-                <Box className={[styles.inputWrapper, tone === 'error' ? styles.error : null]}>
+                <Box
+                    className={[
+                        styles.inputWrapper,
+                        tone === 'error' ? styles.error : null,
+                        variant === 'bordered' ? styles.bordered : null,
+                    ]}
+                >
                     <input {...props} {...extraProps} type={type} ref={ref} />
                 </Box>
             )}